+++ files/default/tests/minitest/hadoop_hdfs_ha_checkconfig_test.rb
@@ -1,8 +1,6 @@
 require File.expand_path('../support/helpers', __FILE__)
 
 describe 'hadoop::hadoop_hdfs_ha_checkconfig' do
-
   include Helpers::Hadoop
   # No tests, since this cookbook just tests attributes
-
 end