+++ files/default/tests/minitest/hbase_checkconfig_test.rb
@@ -1,8 +1,6 @@
 require File.expand_path('../support/helpers', __FILE__)
 
 describe 'hadoop::hbase_checkconfig' do
-
   include Helpers::Hadoop
   # No tests, since this cookbook just tests attributes
-
 end