+++ metadata.rb
@@ -2,9 +2,9 @@ name             'hadoop'
 maintainer       'Cask Data, Inc.'
 maintainer_email 'ops@cask.co'
 license          'Apache 2.0'
-description      'Installs/Configures Hadoop (HDFS/YARN/MRv2), HBase, Hive, Flume, Oozie, Pig, and ZooKeeper'
+description      'Installs/Configures Hadoop (HDFS/YARN/MRv2), HBase, Hive, Flume, Oozie, Pig, Spark, Tez, and ZooKeeper'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.6.1'
+version          '1.9.2'
 
 depends 'yum', '>= 3.0'
 depends 'apt'
@@ -15,6 +15,6 @@ recommends 'java', '~> 1.21'
   supports os
 end
 
-%w(hadoop_cluster hadoop_cluster_rpm hadoop_for_hbase hbase hbase_cluster hive pig zookeeper zookeeper_cluster).each do |cb|
+%w(hadoop_cluster hadoop_cluster_rpm hadoop_for_hbase hbase hbase_cluster hive pig spark tez zookeeper zookeeper_cluster).each do |cb|
   conflicts cb
 end