+++ spec/hive_server2_spec.rb
@@ -7,7 +7,6 @@ describe 'hadoop::hive_server2' do
         node.automatic['domain'] = 'example.com'
         node.default['hive']['hive_site']['hive.support.concurrency'] = 'true'
         node.default['hive']['hive_site']['hive.zookeeper.quorum'] = 'localhost'
-        stub_command('update-alternatives --display hadoop-conf | grep best | awk \'{print $5}\' | grep /etc/hadoop/conf.chef').and_return(false)
         stub_command('update-alternatives --display hive-conf | grep best | awk \'{print $5}\' | grep /etc/hive/conf.chef').and_return(false)
       end.converge(described_recipe)
     end
@@ -31,7 +30,6 @@ describe 'hadoop::hive_server2' do
         node.override['hadoop']['distribution'] = 'cdh'
         node.default['hive']['hive_site']['hive.support.concurrency'] = 'true'
         node.default['hive']['hive_site']['hive.zookeeper.quorum'] = 'localhost'
-        stub_command('update-alternatives --display hadoop-conf | grep best | awk \'{print $5}\' | grep /etc/hadoop/conf.chef').and_return(false)
         stub_command('update-alternatives --display hive-conf | grep best | awk \'{print $5}\' | grep /etc/hive/conf.chef').and_return(false)
       end.converge(described_recipe)
     end