+++ spec/hive_spec.rb
@@ -7,7 +7,7 @@ describe 'hadoop::hive' do
         node.automatic['domain'] = 'example.com'
         node.default['hive']['hive_site']['hive.exec.local.scratchdir'] = '/tmp/hive/scratch'
         node.default['hive']['hive_env']['hive_log_dir'] = '/data/log/hive'
-        stub_command('update-alternatives --display hadoop-conf | grep best | awk \'{print $5}\' | grep /etc/hadoop/conf.chef').and_return(false)
+        stub_command('test -L /var/log/hive').and_return(false)
         stub_command('update-alternatives --display hive-conf | grep best | awk \'{print $5}\' | grep /etc/hive/conf.chef').and_return(false)
       end.converge(described_recipe)
     end
@@ -69,7 +69,6 @@ describe 'hadoop::hive' do
     it 'creates /tmp/hive/scratch directory' do
       expect(chef_run).to create_directory('/tmp/hive/scratch')
     end
-
   end
 
   context 'on Ubuntu 12.04' do
@@ -92,6 +91,5 @@ describe 'hadoop::hive' do
         expect(link).to link_to("/usr/share/java/#{jar}.jar")
       end
     end
-
   end
 end