+++ spec/spark_worker_spec.rb
@@ -1,0 +1,47 @@
+require 'spec_helper'
+
+describe 'hadoop::spark_worker' do
+  context 'on Centos 6.5 x86_64' do
+    let(:chef_run) do
+      ChefSpec::SoloRunner.new(platform: 'centos', version: 6.5) do |node|
+        node.automatic['domain'] = 'example.com'
+        node.default['spark']['release']['install'] = true
+        node.default['spark']['spark_env']['spark_worker_dir'] = '/data/spark/work'
+        stub_command('test -L /var/log/spark').and_return(false)
+        stub_command('update-alternatives --display spark-conf | grep best | awk \'{print $5}\' | grep /etc/spark/conf.chef').and_return(false)
+      end.converge(described_recipe)
+    end
+
+    it 'does not install spark-worker package' do
+      expect(chef_run).not_to install_package('spark-worker')
+    end
+
+    it 'creates /data/spark/work directory' do
+      expect(chef_run).to create_directory('/data/spark/work')
+    end
+  end
+
+  context 'using CDH 5' do
+    let(:chef_run) do
+      ChefSpec::SoloRunner.new(platform: 'centos', version: 6.5) do |node|
+        node.automatic['domain'] = 'example.com'
+        node.override['hadoop']['distribution'] = 'cdh'
+        node.override['hadoop']['distribution_version'] = 5
+        stub_command('test -L /var/log/spark').and_return(false)
+        stub_command('update-alternatives --display spark-conf | grep best | awk \'{print $5}\' | grep /etc/spark/conf.chef').and_return(false)
+      end.converge(described_recipe)
+    end
+
+    it 'installs spark-worker package' do
+      expect(chef_run).to install_package('spark-worker')
+    end
+
+    it 'creates spark-worker service resource, but does not run it' do
+      expect(chef_run).to_not start_service('spark-worker')
+    end
+
+    it 'creates /var/run/spark/work directory' do
+      expect(chef_run).to create_directory('/var/run/spark/work')
+    end
+  end
+end